Detect token invalidations
This adds a few lifecycle management things to make it nicer to user
this. Now, we will automatically delete kernels that are idle, and also
automatically tear down servers that have not been interacted with
recently. Additionally, we periodically check to see if tokens have
expired, and force the user to log in again if they want to spawn a new
server.