Preserve SSH key on server destroy
It's pretty annoying to have to remake your SSH key and re-upload it to
OpenStack every time. Storing the SSH key in the user's home directory
will preserve it across server instantiations, and will really improve
that experience. We generate it at /work/.ssh if it doesn't exist.

Also adds a helper `key_pair_upload` that can automatically upload it
for you.