Add netifaces to built wheels
It looks like somewhere in the dependency chain, a dependency on this
wheel was added. The wheel will fail to build in the second stage due to
not having the GCC tools installed. netifaces is installed somehow by
jupyterhub-chameleon, which is installed in the second stage. This adds
netifaces to the list of wheels explicitly built beforehand to resolve.