Use per-server keypairs for Trovi servers
Named servers / experiment servers have a problem where they end up
splitting the SSH keypairs; a keypair is uploaded and a private key saved,
but then the private key is gone when a new experiment version is created,
but the keypair still has the same name in Nova.

With this change, all named server environments will have their own named
keypair in Nova. This will cause more keypairs to be registered in Nova, but
it should be obvious which ones are Trovi-related and which ones are not.

In the future we can create a mechanism to clean up some of these old keypairs.