Update hub and notebook images
This updates to the latest version of the hub/notebook images.
Some of the start scripts got moved around as there is more/better
functionality in the start-notebook.sh script that invokes them.
In particular the homedir madness is a bit saner. We still do
the symlink from /work => /home/$NB_USER/work but that's about it.